mm: Create FolioFlags
These new functions are the folio analogues of the PageFlags functions.
If CONFIG_DEBUG_VM_PGFLAGS is enabled, we check the folio is not a
tail page at every invocation.  Note that this will also catch the
PagePoisoned case as a poisoned page has every bit set, which would
include PageTail.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>