fs: Convert vfs_dedupe_file_range_compare to folios
Simplify the implementation somewhat by working in pgoff_t instead
of loff_t.  We still only operate on a single page of data at a time
due to using kmap().  A more complex implementation would work on
an entire folio at a time and (if the pages are highmem) map and
unmap, but it's not clear that such a complex implementation would
be worthwhile.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>