fs: Convert ->set_page_dirty to work on folios
set_page_dirty() already passed the head page to the filesystem.
Formalise this by changing the type from page to folio.  Also document
that it returns only true/false by changing the return type to bool.
Add a set_folio_dirty() for callers who already have a folio.  Pass the
mapping pointer to the filesystem as well to avoid any confusion if the
page is truncated during the call to set_page_dirty().

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>